+++ src/main/java/cart/controller/CartController.java
@@ -1,9 +1,9 @@
 package cart.controller;
 
 import cart.authority.Authority;
-import cart.controller.dto.CartItemCreationRequest;
-import cart.controller.dto.CartItemDeleteRequest;
-import cart.controller.dto.MemberIdRequest;
+import cart.controller.dto.request.CartItemCreationRequest;
+import cart.controller.dto.request.CartItemDeleteRequest;
+import cart.controller.dto.request.MemberIdRequest;
 import cart.domain.dto.CartDto;
 import cart.domain.dto.ProductDto;
 import cart.service.CartService;
