+++ src/main/java/cart/controller/ProductController.java
@@ -1,7 +1,7 @@
 package cart.controller;
 
-import cart.controller.dto.ProductCreationRequest;
-import cart.controller.dto.ProductUpdateRequest;
+import cart.controller.dto.request.ProductCreationRequest;
+import cart.controller.dto.request.ProductUpdateRequest;
 import cart.service.AdminService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
