+++ src/main/java/cart/controller/dto/request/CartItemCreationRequest.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.controller.dto.request;
 
 public class CartItemCreationRequest {
     private Long productId;
