+++ src/main/java/cart/controller/dto/request/CartItemDeleteRequest.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.controller.dto.request;
 
 public class CartItemDeleteRequest {
