+++ src/main/java/cart/controller/dto/request/ProductCreationRequest.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.controller.dto.request;
 
 import org.hibernate.validator.constraints.Range;
