+++ src/main/java/cart/controller/dto/request/ProductUpdateRequest.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.controller.dto.request;
 
 import org.hibernate.validator.constraints.Range;
