+++ src/main/java/cart/controller/dto/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.controller.dto.response;
 
 import cart.domain.dto.ProductDto;
